refactor: updated structure and moved settings.py and fomod_dialog.py respectively to src/gui/app_views and src/gui/dashboard_views
